requirements.txt
Django==2.2.16
django-crispy-forms==1.9.2
django-environ==0.4.5
django-filter==2.4.0
pytz==2020.1
sqlparse==0.3.1
